Move mapreduce() to separate file
