+++ madoop/exceptions.py
@@ -1,0 +1,8 @@
+"""Madoop Exception Types.
+
+Andrew DeOrio <awdeorio@umich.edu>
+
+"""
+
+class MadoopError(Exception):
+    """Top level exception raised by Madoop functions."""
